requirements.txt
Flask==2.3.3
gunicorn==21.2.0
google-cloud-firestore==2.11.1
requests==2.31.0
